requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.2
django-crispy-forms==2.4
django-environ==0.12.0
django-parler==2.3
django-rosetta==0.10.2
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
polib==1.2.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
